CRUD funcionando. Script DDBB configurado para MySql. Comandos docker para ejecutar contenedor mysql. Exception Handler generico

+++ src/main/java/iaew/tpi/apiClientes/transferObjects/ResponseWrapperDto.java
@@ -2,7 +2,7 @@ package iaew.tpi.apiClientes.transferObjects;
 
 import java.util.List;
 
-public record ResponseWrapperDto(int page, int size, int count, List<ClienteDto> results) {
+public record ResponseWrapperDto(int page, int size, long count, List<ClienteDto> results) {
 
 
 }
